refactor(navigation-secondary): refactor overlay
