+++ elements/rh-navigation-secondary/rh-navigation-secondary-overlay.ts
@@ -1,20 +1,8 @@
-import { html, LitElement } from 'lit';
+import { LitElement } from 'lit';
 import { customElement } from 'lit/decorators/custom-element.js';
-import { state } from 'lit/decorators/state.js';
-
-import { ComposedEvent } from '@patternfly/pfe-core';
-import { observed } from '@patternfly/pfe-core/decorators.js';
+import { property } from 'lit/decorators/property.js';
 
 import styles from './rh-navigation-secondary-overlay.css';
-
-export class SecondaryNavOverlayChangeEvent extends ComposedEvent {
-  constructor(
-    public open: boolean,
-    public toggle: HTMLElement
-  ) {
-    super('overlay-change');
-  }
-}
 
 /**
  * @summary An overlay element to cover content with an opacity when navigation is expanded.
@@ -23,16 +11,7 @@ export class SecondaryNavOverlayChangeEvent extends ComposedEvent {
 export class RhNavigationSecondaryOverlay extends LitElement {
   static readonly styles = [styles];
 
-  @observed
-  @state() open = false;
-
-  render() {
-    return html``;
-  }
-
-  protected _openChanged(_oldValue?: boolean, newValue?: boolean) {
-    this.toggleAttribute('open', newValue);
-  }
+  @property({ type: Boolean, reflect: true }) open = false;
 }
 
 declare global {
